Move testing modules under src/node
