--- packages/testing/src/index.ts
@@ -1,3 +1,0 @@
-export * as temporary from './temporary'
-export * as workspace from './workspace'
-export * as models from './models'
